Rename test-d tests to allow JSX syntax
