+++ packages/liveblocks-react/test-d/factories.test-d.tsx
@@ -1,3 +1,6 @@
+import React from "react";
+React; // To silence tsd warning
+
 import type {
   Json,
   LiveList,
